% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalTests.R
\name{.asHtest}
\alias{.asHtest}
\title{Standardize statistical inference results to R's htest convention.}
\usage{
.asHtest(result, method, data.name)
}
\description{
Standardize statistical inference results to R's htest convention.
}
\keyword{internal}
