% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilistic.R
\name{expectedRewardsBeforeHittingA}
\alias{expectedRewardsBeforeHittingA}
\title{Expected first passage Rewards for a set of states in a markovchain}
\usage{
expectedRewardsBeforeHittingA(markovchain, A, state, rewards, n)
}
\arguments{
\item{markovchain}{the markovchain-class object}

\item{A}{set of states for first passage expected reward}

\item{state}{initial state}

\item{rewards}{vector depicting rewards coressponding to states}

\item{n}{no of steps of the process}
}
\value{
returns a expected reward (numerical value) as described above
}
\description{
Given a markovchain object and reward values for every state,
function calculates expected reward value for a set A of states after n 
steps.
}
\details{
The function returns the value of expected first passage 
rewards given rewards coressponding to every state, an initial state
and number of steps.
}
\author{
Sai Bhargav Yalamanchi, Vandit Jain
}
