% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_probabilistic.R
\name{expectedRewardsforA}
\alias{expectedRewardsforA}
\title{Expected first passage Rewards for a set of states in a markovchain}
\usage{
expectedRewardsforA(markovchain, A, state, rewards, n)
}
\arguments{
\item{markovchain}{the markovchain-class object}

\item{A}{set of states for first passage expected reward}

\item{state}{initial state}

\item{rewards}{vector depicting rewards coressponding to states}

\item{n}{no of steps of the process}
}
\value{
returns a expected reward (numerical value) as described above
}
\description{
The function returns the value of expected first passage 
rewards given rewards coressponding to every state, an initial state
}
\details{
The function returns the value of expected first passage 
rewards given rewards coressponding to every state, an initial state
}
\author{
Sai Bhargav Yalamanchi, Vandit Jain
}
