% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_fommc.R
\name{fitFirstOrderMultivarMC}
\alias{fitFirstOrderMultivarMC}
\title{Function to fit First Order Multivariate Markovchain}
\usage{
fitFirstOrderMultivarMC(seqMat)
}
\arguments{
\item{seqMat}{Matrix whose each row is a sequence}
}
\value{
A list of two parameters
        1. lambdas
        2. transition matrices
}
\description{
Given a number of sequence(in the form of matrix) this function will fit the 
             First Order Multivariate Markovchain
}
\examples{
smat <- matrix(c('a','a','b','b','a','c','a',
                 'b','a','b','c','c','b','a',
                 'a','b','b','a','c','a','b',
                 'c','c','c','c','a','a','a'),
               nrow = 4, byrow = TRUE)
               
fommcFit <- fitFirstOrderMultivarMC(smat)

}
\author{
Giorgio Spedicato, Deepak Yadav
}
\references{
W.-K. Ching et al. / Linear Algebra and its Applications
}
\seealso{
\code{\link{markovchainFit}}
}

