% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmcProbabilisticSAI.R
\name{freq2Generator}
\alias{freq2Generator}
\title{Returns a generator matrix corresponding to frequency matrix}
\usage{
freq2Generator(P, t = 1, method = "QO", logmethod = "Eigen")
}
\arguments{
\item{P}{frequency matrix}

\item{t}{(default value = 1)}

\item{method}{one among "QO"(Quasi optimaisation), "WA"(weighted adjustment), "DA"(diagonal adjustment)}

\item{logmethod}{method for computation of matrx algorithm (by default : Eigen)}
}
\value{
returns a generator matix with same dimnames
}
\description{
The function provides interface to calculate generator matrix corresponding to 
a frequency matrix and time taken
}
