% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_classesAndMethods.R
\docType{methods}
\name{name}
\alias{name}
\alias{name,markovchain-method}
\title{Method to retrieve name of markovchain object}
\usage{
name(object)

\S4method{name}{markovchain}(object)
}
\arguments{
\item{object}{A markovchain object}
}
\description{
This method returns the name of markovchain object
}
\examples{
statesNames <- c("a", "b", "c")
markovB <- new("markovchain", states = statesNames, transitionMatrix =
                matrix(c(0.2, 0.5, 0.3, 0, 1, 0, 0.1, 0.8, 0.1), nrow = 3,
                byrow = TRUE, dimnames=list(statesNames,statesNames)),
               name = "A markovchain Object" 
)
name(markovB)

}
\author{
Giorgio Spedicato, Deepak Yadav
}
