% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHigherOrder.R
\docType{class}
\name{HigherOrderMarkovChain-class}
\alias{HigherOrderMarkovChain-class}
\title{Higher order Markov Chains class}
\description{
The S4 class that describes \code{HigherOrderMarkovChain} objects.
}
