% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_hommc.R
\docType{methods}
\name{show,hommc-method}
\alias{show,hommc-method}
\title{Function to display the details of hommc object}
\usage{
\S4method{show}{hommc}(object)
}
\arguments{
\item{object}{An object of class hommc}
}
\description{
This is a convenience function to display the slots of hommc object
             in proper format
}
