% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmcClassesAndMethods.R
\docType{class}
\name{ictmc-class}
\alias{ictmc-class}
\alias{ictmc}
\title{An S4 class for representing Imprecise Continuous Time Markovchains}
\description{
An S4 class for representing Imprecise Continuous Time Markovchains
}
\section{Slots}{

\describe{
\item{\code{states}}{a vector of states present in the ICTMC model}

\item{\code{Q}}{matrix representing the generator demonstrated in the form of variables}

\item{\code{range}}{a matrix that stores values of range of variables}

\item{\code{name}}{name given to ICTMC}
}}

