% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmcProbabilisticSAI.R
\name{impreciseProbabilityatT}
\alias{impreciseProbabilityatT}
\title{Calculating full conditional probability using lower rate transition matrix}
\usage{
impreciseProbabilityatT(C,i,t,s,error)
}
\arguments{
\item{C}{a ictmc class object}

\item{i}{initial state at time t}

\item{t}{initial time t. Default value = 0}

\item{s}{final time}

\item{error}{error rate. Default value = 0.001}
}
\description{
This function calculates full conditional probability at given 
time s using lower rate transition matrix
}
