% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kullback}
\alias{kullback}
\title{Example from Kullback and Kupperman Tests for Contingency Tables}
\format{A list containing two 6x6 non - negative integer matrices}
\usage{
data(kullback)
}
\description{
A list of two matrices representing raw transitions between two states
}
\keyword{datasets}
