% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_functions4Fitting.R
\name{markovchainSequenceParallel}
\alias{markovchainSequenceParallel}
\title{Function to generate a list of sequence of states in parallel from non-homogeneous Markov chains.}
\usage{
markovchainSequenceParallel(n, object,
  t0 = sample(object@markovchains[[1]]@states, 1), num.cores = NULL)
}
\arguments{
\item{n}{Sample size}

\item{object}{markovchainList object}

\item{t0}{Initial state}

\item{num.cores}{Number of cores}
}
\description{
Provided any markovchainList object, it returns a list of sequence of states coming 
from the underlying stationary distribution.
}

