% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{markovchainSequenceParallelRcpp}
\alias{markovchainSequenceParallelRcpp}
\title{Function to generate a list of sequence of states in parallel from non-homogeneous Markov chains.}
\usage{
markovchainSequenceParallelRcpp(listObject, n, include_t0 = FALSE)
}
\arguments{
\item{listObject}{markovchainList object}

\item{n}{Sample size}

\item{include_t0}{Specify if the initial state shall be used}
}
\value{
A List
}
\description{
Provided any markovchainList object, it returns a list of sequence of states coming 
from the underlying stationary distribution.
}
\examples{
statesNames <- c("a")
mcA <- new("markovchain", states = statesNames, transitionMatrix = 
       matrix(c(1), nrow = 1, byrow = TRUE, 
       dimnames = list(statesNames, statesNames)))
 
statesNames <- c("a","b")
mcB <- new("markovchain", states = statesNames, transitionMatrix = 
       matrix(c(0.5, 0.5, 0.3, 0.7), nrow = 2, byrow = TRUE, 
       dimnames = list(statesNames, statesNames)))
       
statesNames <- c("a","b","c")       
mcC <- new("markovchain", states = statesNames, transitionMatrix = 
       matrix(c(0.2, 0.5, 0.3, 0, 0.2, 0.8, 0.1, 0.8, 0.1), nrow = 3, 
       byrow = TRUE, dimnames = list(statesNames, statesNames)))  

mclist <- new("markovchainList", markovchains = list(mcA, mcB, mcC))   

markovchainSequenceParallelRcpp(mclist, 99999, TRUE)


}
\author{
Giorgio Spedicato, Deepak Yadav
}

