% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesAndMethods.R
\name{names,markovchain-method}
\alias{names,markovchain-method}
\title{Returns the states for a Markov chain object}
\usage{
\S4method{names}{markovchain}(x)
}
\arguments{
\item{x}{object we want to return states for}
}
\description{
Returns the states for a Markov chain object
}
