% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_classesAndMethods.R
\docType{methods}
\name{name<-}
\alias{name<-}
\alias{name<-,markovchain-method}
\title{Method to set name of markovchain object}
\usage{
name(object) <- value

\S4method{name}{markovchain}(object) <- value
}
\arguments{
\item{object}{A markovchain object}

\item{value}{New name of markovchain object}
}
\description{
This method modifies the existing name of markovchain object
}
\examples{
statesNames <- c("a", "b", "c")
markovB <- new("markovchain", states = statesNames, transitionMatrix =
                matrix(c(0.2, 0.5, 0.3, 0, 1, 0, 0.1, 0.8, 0.1), nrow = 3,
                byrow = TRUE, dimnames=list(statesNames,statesNames)),
                name = "A markovchain Object" 
)
name(markovB) <- "dangerous mc"

}
\author{
Giorgio Spedicato, Deepak Yadav
}
