% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tm_abs}
\alias{tm_abs}
\title{Single Year Corporate Credit Rating Transititions}
\format{The format is:
num [1:8, 1:8] 17 2 0 0 0 0 0 0 1 455 ...
- attr(*, "dimnames")=List of 2
..$ : chr [1:8] "AAA" "AA" "A" "BBB" ...
..$ : chr [1:8] "AAA" "AA" "A" "BBB" ...}
\usage{
data(tm_abs)
}
\description{
Matrix of Standard and Poor's Global Corporate Rating Transition Frequencies 2000 (NR Removed)
}
\examples{
data(tm_abs)
}
\references{
European Securities and Markets Authority, 2016 
https://cerep.esma.europa.eu/cerep-web/statistics/transitionMatrice.xhtml
}
\keyword{datasets}
