% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_statisticalTests.R
\name{verifyHomogeneity}
\alias{verifyHomogeneity}
\title{Verify Homogeneity across transition matrices}
\usage{
verifyHomogeneity(inputList, verbose = TRUE)
}
\arguments{
\item{inputList}{A list of items that can coerced to transition matrices}

\item{verbose}{Does output should be printed out}
}
\value{
a list of transition matrices?
}
\description{
Verifies that the s elements in the input list belongs to the same DTMC
}
\examples{

data(kullback)
verifyHomogeneity(inputList=kullback,verbose=TRUE)

}
