% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/classesAndMethods.R,
%   R/probabilistic.R
\name{period}
\alias{period}
\alias{absorbingStates}
\alias{transientStates}
\alias{recurrentStates}
\alias{canonicForm}
\alias{communicatingClasses}
\alias{recurrentClasses}
\title{Various function to perform structural analysis of DTMC}
\usage{
period(object)

absorbingStates(object)

transientStates(object)

recurrentStates(object)

canonicForm(object)

communicatingClasses(object)

recurrentClasses(object)
}
\arguments{
\item{object}{A \code{markovchain} object.}
}
\value{
vector, matrix or list
}
\description{
These functions return absorbing and transient states of the \code{markovchain} objects.
}
\examples{
statesNames <- c("a", "b", "c")
markovB <- new("markovchain", states = statesNames, transitionMatrix =
                   matrix(c(0.2, 0.5, 0.3,
                              0,   1,   0,
                            0.1, 0.8, 0.1), nrow = 3, byrow = TRUE, 
                            dimnames = list(statesNames, statesNames)
              ))
              
communicatingClasses(markovB)               
recurrentClasses(markovB)
absorbingStates(markovB)
transientStates(markovB)
canonicForm(markovB)

# periodicity analysis : 1
E <- matrix(c(0, 1, 0, 0, 0.5, 0, 0.5, 0, 0, 0.5, 0, 0.5, 0, 0, 1, 0), 
            nrow = 4, ncol = 4, byrow = TRUE)
mcE <- new("markovchain", states = c("a", "b", "c", "d"), 
          transitionMatrix = E, 
          name = "E")

is.irreducible(mcE) #true
period(mcE) #2

# periodicity analysis : 2
myMatr <- matrix(c(0, 0, 1/2, 1/4, 1/4, 0, 0,
                   0, 0, 1/3, 0, 2/3, 0, 0,
                   0, 0, 0, 0, 0, 1/3, 2/3,
                   0, 0, 0, 0, 0, 1/2, 1/2,
                   0, 0, 0, 0, 0, 3/4, 1/4,
                   1/2, 1/2, 0, 0, 0, 0, 0,
                   1/4, 3/4, 0, 0, 0, 0, 0), byrow = TRUE, ncol = 7)
myMc <- new("markovchain", transitionMatrix = myMatr)
period(myMc)

}
\references{
Feres, Matlab listing for markov chain.
}
\seealso{
\code{\linkS4class{markovchain}}
}
\author{
Giorgio Alfredo Spedicato
}
