% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_probabilistic.R
\name{expectedRewardsforA}
\alias{expectedRewardsforA}
\title{Expected first passage Rewards for a set of states in a markovchain}
\usage{
expectedRewardsforA(markovchain, A, state, rewards, n)
}
\description{
@description The function returns the value of expected first passage 
 rewards given rewards coressponding to every state, an initial state
 
 @usage expectedRewards(markovchain,n,rewards)
 
 @param markovchain the markovchain-class object
 @param A set of states for first passage expected reward
 @param state initial state
 @param rewards vector depicting rewards coressponding to states
 @param n no of steps of the process
 
 @details The function returns the value of expected first passage 
 rewards given rewards coressponding to every state, an initial state
 
 @return returns a expected reward (numerical value) as described above
 
 @author Sai Bhargav Yalamanchi, Vandit Jain
 
 @export
}
