% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ctmcProbabilisticSAI.R
\name{ExpectedTime}
\alias{ExpectedTime}
\title{Return expected hitting time from state i to state j}
\usage{
ExpectedTime(C, i, j)
}
\arguments{
\item{C}{A CTMC object}

\item{i}{initial state i}

\item{j}{final state j}
}
\value{
a value that returns expected hitting times from i to j
}
\description{
Return expected hitting time from state i to state j
}

