% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_classesAndMethods.R
\docType{methods}
\name{name}
\alias{name}
\alias{name,markovchain-method}
\title{Method to retrieve name of markovchain object}
\usage{
name(object)

\S4method{name}{markovchain}(object)
}
\arguments{
\item{object}{A markovchain object}
}
\description{
This method returns the name of markovchain object
}

