% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilistic.R
\name{is.stochasticallyMonotone}
\alias{is.stochasticallyMonotone}
\title{Check if a Markov chain is stochastically monotone}
\usage{
is.stochasticallyMonotone(object)
}
\arguments{
\item{object}{A \code{markovchain} object or a transition matrix.}
}
\value{
A boolean value.
}
\description{
Verifies whether a transition matrix is stochastically monotone.
}
\examples{
P <- matrix(c(0.8, 0.2,
              0.3, 0.7), nrow = 2, byrow = TRUE)
is.stochasticallyMonotone(P)

}
