% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilistic.R
\name{is.stochasticallyMonotone}
\alias{is.stochasticallyMonotone}
\alias{is.stochasticallyMonotone,markovchain-method}
\alias{is.stochasticallyMonotone,matrix-method}
\alias{is.stochasticallyMonotone,ANY-method}
\title{Check if a Markov chain is stochastically monotone}
\usage{
is.stochasticallyMonotone(object)

\S4method{is.stochasticallyMonotone}{markovchain}(object)

\S4method{is.stochasticallyMonotone}{matrix}(object)

\S4method{is.stochasticallyMonotone}{ANY}(object)
}
\arguments{
\item{object}{A markovchain object or a transition matrix.}
}
\value{
A boolean value.
}
\description{
Verifies if the transition matrix of the Markov chain is stochastically monotone.
}
