% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_statisticalTests.R, R/data.R
\docType{data}
\name{kullback}
\alias{kullback}
\title{Example from Kullback and Kupperman Tests for Contingency Tables}
\format{A list containing two 6x6 non - negative integer matrices}
\source{
Kullback, Kupperman Tests for Contingency Tables Technometrics
}
\usage{
kullback

data(kullback)
}
\description{
A list of two matrices representing raw transitions between two states

A list of two matrices representing raw transitions between two states
}
\keyword{datasets}

