% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_classesAndMethods.R
\docType{class}
\name{markovchain-class}
\alias{markovchain-class}
\alias{*}
\alias{[}
\alias{^}
\alias{==}
\alias{!=}
\alias{coerce}
\alias{dim}
\alias{plot}
\alias{print}
\alias{show}
\alias{summary}
\alias{t}
\title{Markov chain class}
\description{
The S4 class that describes \code{markovchain} objects.
}
