% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovchain.R
\docType{package}
\name{markovchain}
\alias{markovchain}
\alias{markovchain-package}
\title{markovchain: A package for Markov chains}
\description{
Package to represent DTMC and CTMC Markov chains, extract information from
them, related to their states, their stationarity and do inference
}
