% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_functions4Fitting.R
\name{noofVisitsDist}
\alias{noofVisitsDist}
\title{return a joint pdf of the number of visits to the various states of the DTMC}
\usage{
noofVisitsDist(markovchain,N,state)
}
\arguments{
\item{markovchain}{a markovchain-class object}

\item{N}{no of steps}

\item{state}{the initial state}
}
\value{
a numeric vector depicting the above described probability density function.
}
\description{
This function would return a joint pdf of the number of visits to
the various states of the DTMC during the first N steps.
}
\details{
This function would return a joint pdf of the number of visits to
the various states of the DTMC during the first N steps.
}
\examples{
transMatr<-matrix(c(0.4,0.6,.3,.7),nrow=2,byrow=TRUE)
simpleMc<-new("markovchain", states=c("a","b"),
             transitionMatrix=transMatr, 
             name="simpleMc")   
noofVisitsDist(simpleMc,5,"a")

}
\author{
Vandit Jain
}
