% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rain}
\alias{rain}
\title{Alofi island daily rainfall}
\format{A data frame with 1096 observations on the following 2 variables.
\describe{
  \item{\code{V1}}{a numeric vector, showing original coding}
  \item{\code{rain}}{a character vector, showing daily rainfall millilitres brackets}
}}
\source{
Avery Henderson
}
\usage{
data(rain)
}
\description{
Rainfall measured in Alofi Island
}
\examples{
data(rain)
rainMc<-markovchainFit(data=rain$rain)
}
\references{
Avery Henderson, Fitting markov chain models on discrete time series such as DNA sequences
}
\keyword{datasets}
