% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_classesAndMethods.R
\docType{methods}
\name{name<-}
\alias{name<-}
\alias{name<-,markovchain-method}
\title{Method to set name of markovchain object}
\usage{
name(object) <- value

\S4method{name}{markovchain}(object) <- value
}
\arguments{
\item{object}{A markovchain object}

\item{value}{New name of markovchain object}
}
\description{
This method modify the existing name of markovchain object
}

